Rename Time to TimeExtras
